+++ packages/soda-react/CHANGELOG.md
@@ -1,5 +1,12 @@
 # soda-react
 
+## 0.0.5
+
+### Patch Changes
+
+- Updated dependencies
+    - deepsea-tools@5.34.0
+
 ## 0.0.4
 
 ### Patch Changes